refactor(pill-next): Tag composes PillNext instead of duplicating it
Tag now wraps PillNext and passes badge content via the children prop,
instead of duplicating all the pill logic. This ensures both components
stay in sync when PillNext is updated.

- Add children prop to PillNext (renders between label and remove button)
- Move Tag into its own src/Tag/ subdirectory
- Remove Tag.styles.ts (badge spacing is the only custom style)
