--- packages/components/pill-next/src/Tag.styles.ts
@@ -1,89 +1,0 @@
-import { css } from '@emotion/css';
-import tokens from '@contentful/f36-tokens';
-import type { PillNextVariant } from './PillNext.types';
-
-const variantStyles: Record<
-  PillNextVariant,
-  { background: string; border: string }
-> = {
-  secondary: {
-    background: tokens.gray100,
-    border: tokens.gray200,
-  },
-  primary: {
-    background: tokens.blue100,
-    border: tokens.blue200,
-  },
-  warning: {
-    background: tokens.orange100,
-    border: tokens.orange200,
-  },
-  negative: {
-    background: tokens.red100,
-    border: tokens.red200,
-  },
-};
-
-export function getTagStyles(
-  variant: PillNextVariant,
-  hasRemoveButton: boolean,
-) {
-  const { background, border } = variantStyles[variant];
-
-  return {
-    tag: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      borderRadius: '16px',
-      height: '32px',
-      paddingTop: tokens.spacing2Xs,
-      paddingBottom: tokens.spacing2Xs,
-      paddingLeft: tokens.spacingS,
-      paddingRight: hasRemoveButton ? tokens.spacing2Xs : tokens.spacingS,
-      border: `1px solid ${border}`,
-      backgroundColor: background,
-      fontFamily: tokens.fontStackPrimary,
-      boxSizing: 'border-box',
-    }),
-    leadingIconWrapper: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      flexShrink: 0,
-      marginRight: tokens.spacing2Xs,
-    }),
-    leadingIcon: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-    }),
-    label: css({
-      fontSize: tokens.fontSizeM,
-      fontWeight: tokens.fontWeightMedium,
-      lineHeight: tokens.lineHeightM,
-      whiteSpace: 'nowrap',
-    }),
-    badge: css({
-      display: 'inline-flex',
-      alignItems: 'center',
-      marginLeft: tokens.spacing2Xs,
-    }),
-    removeButton: css({
-      '&&': {
-        width: '24px',
-        height: '24px',
-        minHeight: 'auto',
-        minWidth: 'auto',
-        padding: tokens.spacing2Xs,
-        borderRadius: '50%',
-        marginLeft: tokens.spacingXs,
-        mixBlendMode: 'luminosity',
-      },
-      '&&:hover:not(:disabled)': {
-        backgroundColor: tokens.gray300,
-      },
-      '&&:disabled': {
-        backgroundColor: 'transparent',
-        color: tokens.gray400,
-      },
-    }),
-  };
-}
