+++ packages/components/pill-next/src/Tag/index.ts
@@ -1,0 +1,2 @@
+export { Tag } from './Tag';
+export type { TagProps, TagInternalProps } from './Tag';
